QML Architecture Updates (#145)
* Updated structural changes

* Modified docstrings

---------
