+++ src/modules/applications/qml/generative_modeling/circuits/CircuitCardinality.py
@@ -13,28 +13,26 @@
 #  limitations under the License.
 
 from typing import Union, TypedDict
-from itertools import combinations
-from scipy.special import binom
 
-from modules.circuits.Circuit import Circuit
+from modules.applications.qml.generative_modeling.circuits.CircuitGenerative import CircuitGenerative
 from modules.applications.qml.generative_modeling.mappings.LibraryQiskit import LibraryQiskit
 from modules.applications.qml.generative_modeling.mappings.LibraryPennylane import LibraryPennylane
 from modules.applications.qml.generative_modeling.mappings.PresetQiskitNoisyBackend import PresetQiskitNoisyBackend
 from modules.applications.qml.generative_modeling.mappings.CustomQiskitNoisyBackend import CustomQiskitNoisyBackend
 
 
-class CircuitCopula(Circuit):
+class CircuitCardinality(CircuitGenerative):
     """
     This class generates a library-agnostic gate sequence, i.e. a list containing information
-    about the gates and the wires they act on. The marginal distributions generated by the copula
-    are uniformaly distributed.
+    about the gates and the wires they act on.
+    The circuit follows the implementation by Gili et al. https://arxiv.org/abs/2207.13645
     """
 
     def __init__(self):
         """
         Constructor method.
         """
-        super().__init__("DiscreteCopula")
+        super().__init__("CircuitCardinality")
         self.submodule_options = [
             "LibraryQiskit",
             "LibraryPennylane",
@@ -42,33 +40,24 @@ class CircuitCopula(Circuit):
             "PresetQiskitNoisyBackend"
         ]
 
-    @staticmethod
-    def get_requirements() -> list[dict]:
-        """
-        Returns requirements of this module.
-
-        :return: List of dict with requirements of this module
-        """
-        return [{"name": "scipy", "version": "1.12.0"}]
-
     def get_parameter_options(self) -> dict:
         """
-        Returns the configurable settings for this Copula Circuit.
+        Returns the configurable settings for this circuit.
 
-        :return: Dictionary of parameter options
+        :return: Dictionary with parameter options
         .. code-block:: python
 
             return {
-                "depth": {
-                    "values": [1, 2, 3, 4, 5],
-                    "description": "What depth do you want?"
-                },
-            }
-
+                    "depth": {
+                        "values": [2, 4, 8, 16],
+                        "description": "What depth do you want?"
+                    }
+                }
         """
         return {
+
             "depth": {
-                "values": [1, 2, 3, 4, 5],
+                "values": [2, 4, 8, 16],
                 "description": "What depth do you want?"
             },
         }
@@ -76,10 +65,10 @@ class CircuitCopula(Circuit):
     def get_default_submodule(self, option: str) \
             -> Union[LibraryQiskit, LibraryPennylane, PresetQiskitNoisyBackend, CustomQiskitNoisyBackend]:
         """
-        Returns the default submodule based on the given option.
+        Returns the default submodule based on the provided option.
 
-        :param option: The submodule option to select
-        :return: Instance of the selected submodule
+        :param option: The name of the submodule
+        :return: Instance of the default submodule
         :raises NotImplemented: If the provided option is not implemented
         """
         if option == "LibraryQiskit":
@@ -106,42 +95,37 @@ class CircuitCopula(Circuit):
 
     def generate_gate_sequence(self, input_data: dict, config: Config) -> dict:
         """
-        Returns gate sequence of copula architecture.
+        Returns gate sequence of cardinality circuit architecture.
 
         :param input_data: Collection of information from the previous modules
         :param config: Config specifying the number of qubits of the circuit
-        :return: Dictionary including the gate sequence of the Copula Circuit
+        :return: Dictionary including the gate sequence of the Cardinality Circuit
         """
-        n_registers = input_data["n_registers"]
         n_qubits = input_data["n_qubits"]
-        depth = config["depth"]
-        n = n_qubits // n_registers
+        depth = config["depth"] // 2
 
         gate_sequence = []
-        for k in range(n):
-            gate_sequence.append(["Hadamard", [k]])
 
-        for j in range(n_registers - 1):
-            for k in range(n):
-                gate_sequence.append(["CNOT", [k, k + n * (j + 1)]])
+        for k in range(n_qubits):
+            gate_sequence.append(["RX", [k]])
+            gate_sequence.append(["RZ", [k]])
 
-        gate_sequence.append(["Barrier", None])
+        for d in range(depth):
+            gate_sequence.append(["Barrier", None])
+            for k in range(n_qubits - 1):
+                gate_sequence.append(["RXX", [k, k + 1]])
+            gate_sequence.append(["Barrier", None])
 
-        shift = 0
-        for _ in range(depth):
-            for k in range(n):
-                for j in range(n_registers):
-                    gate_sequence.append(["RZ", [j * n + k]])
-                    gate_sequence.append(["RX", [j * n + k]])
-                    gate_sequence.append(["RZ", [j * n + k]])
+            if d == depth - 2:
+                for k in range(n_qubits):
+                    gate_sequence.append(["RX", [k]])
+                    gate_sequence.append(["RZ", [k]])
+                    gate_sequence.append(["RX", [k]])
 
-            k = 3 * n + shift
-            for i, j in combinations(range(n), 2):
-                for l in range(n_registers):
-                    gate_sequence.append(["RXX", [l * n + i, l * n + j]])
-
-                k += 1
-            shift += 3 * n + int(binom(n, 2))
+            elif d < depth - 2:
+                for k in range(n_qubits):
+                    gate_sequence.append(["RX", [k]])
+                    gate_sequence.append(["RZ", [k]])
 
         gate_sequence.append(["Barrier", None])
 
@@ -150,9 +134,9 @@ class CircuitCopula(Circuit):
 
         output_dict = {
             "gate_sequence": gate_sequence,
-            "circuit_name": "Copula",
+            "circuit_name": "Cardinality",
             "n_qubits": n_qubits,
-            "n_registers": n_registers,
+            "n_registers": input_data["n_registers"],
             "depth": depth,
             "histogram_train": input_data["histogram_train"],
             "store_dir_iter": input_data["store_dir_iter"],
