+++ src/modules/applications/qml/generative_modeling/circuits/CircuitCopula.py
@@ -13,25 +13,28 @@
 #  limitations under the License.
 
 from typing import Union, TypedDict
+from itertools import combinations
+from scipy.special import binom
 
-from modules.circuits.Circuit import Circuit
+from modules.applications.qml.generative_modeling.circuits.CircuitGenerative import CircuitGenerative
 from modules.applications.qml.generative_modeling.mappings.LibraryQiskit import LibraryQiskit
 from modules.applications.qml.generative_modeling.mappings.LibraryPennylane import LibraryPennylane
 from modules.applications.qml.generative_modeling.mappings.PresetQiskitNoisyBackend import PresetQiskitNoisyBackend
 from modules.applications.qml.generative_modeling.mappings.CustomQiskitNoisyBackend import CustomQiskitNoisyBackend
 
 
-class CircuitStandard(Circuit):
+class CircuitCopula(CircuitGenerative):
     """
     This class generates a library-agnostic gate sequence, i.e. a list containing information
-    about the gates and the wires they act on.
+    about the gates and the wires they act on. The marginal distributions generated by the copula
+    are uniformaly distributed.
     """
 
     def __init__(self):
         """
         Constructor method.
         """
-        super().__init__("DiscreteStandard")
+        super().__init__("DiscreteCopula")
         self.submodule_options = [
             "LibraryQiskit",
             "LibraryPennylane",
@@ -46,29 +49,28 @@ class CircuitStandard(Circuit):
 
         :return: List of dict with requirements of this module
         """
-        return []
+        return [{"name": "scipy", "version": "1.12.0"}]
 
     def get_parameter_options(self) -> dict:
         """
-        Returns the configurable settings for this standard circuit.
+        Returns the configurable settings for this Copula Circuit.
 
-        :return: Dictionary of parameter options.
+        :return: Dictionary of parameter options
         .. code-block:: python
 
             return {
-                    "depth": {
-                        "values": [1, 2, 3],
-                        "description": "What depth do you want?"
-                    }
-                }
+                "depth": {
+                    "values": [1, 2, 3, 4, 5],
+                    "description": "What depth do you want?"
+                },
+            }
 
         """
         return {
-
             "depth": {
-                "values": [1, 2, 3],
+                "values": [1, 2, 3, 4, 5],
                 "description": "What depth do you want?"
-            }
+            },
         }
 
     def get_default_submodule(self, option: str) \
@@ -104,36 +106,51 @@ class CircuitStandard(Circuit):
 
     def generate_gate_sequence(self, input_data: dict, config: Config) -> dict:
         """
-        Returns gate sequence of standard architecture.
+        Returns gate sequence of copula architecture.
 
         :param input_data: Collection of information from the previous modules
         :param config: Config specifying the number of qubits of the circuit
-        :return: Dictionary including the gate sequence of the Standard Circuit
+        :return: Dictionary including the gate sequence of the Copula Circuit
         """
         n_registers = input_data["n_registers"]
         n_qubits = input_data["n_qubits"]
         depth = config["depth"]
+        n = n_qubits // n_registers
 
         gate_sequence = []
+        for k in range(n):
+            gate_sequence.append(["Hadamard", [k]])
 
+        for j in range(n_registers - 1):
+            for k in range(n):
+                gate_sequence.append(["CNOT", [k, k + n * (j + 1)]])
+
+        gate_sequence.append(["Barrier", None])
+
+        shift = 0
         for _ in range(depth):
-            for k in range(n_qubits):
-                gate_sequence.append(["RY", [k]])
+            for k in range(n):
+                for j in range(n_registers):
+                    gate_sequence.append(["RZ", [j * n + k]])
+                    gate_sequence.append(["RX", [j * n + k]])
+                    gate_sequence.append(["RZ", [j * n + k]])
 
-            for k in range(n_qubits - 1):
-                gate_sequence.append(["RYY", [k, k + 1]])
+            k = 3 * n + shift
+            for i, j in combinations(range(n), 2):
+                for l in range(n_registers):
+                    gate_sequence.append(["RXX", [l * n + i, l * n + j]])
 
-            for k in range(n_qubits - 1):
-                gate_sequence.append(["CRY", [k, k + 1]])
+                k += 1
+            shift += 3 * n + int(binom(n, 2))
 
-            gate_sequence.append(["Barrier", None])
+        gate_sequence.append(["Barrier", None])
 
         for k in range(n_qubits):
             gate_sequence.append(["Measure", [k, k]])
 
         output_dict = {
             "gate_sequence": gate_sequence,
-            "circuit_name": "Standard",
+            "circuit_name": "Copula",
             "n_qubits": n_qubits,
             "n_registers": n_registers,
             "depth": depth,
