+++ src/modules/applications/qml/generative_modeling/mappings/CustomQiskitNoisyBackend.py
@@ -26,9 +26,9 @@ from qiskit.providers import Backend
 from qiskit_aer import Aer, AerSimulator, noise
 from qiskit_aer.noise import NoiseModel
 
-from modules.training.QCBM import QCBM
-from modules.training.Inference import Inference
-from modules.applications.qml.generative_modeling.mappings.Library import Library
+from modules.applications.qml.generative_modeling.training.QCBM import QCBM
+from modules.applications.qml.generative_modeling.training.Inference import Inference
+from modules.applications.qml.generative_modeling.mappings.LibraryGenerative import LibraryGenerative
 
 logging.getLogger("NoisyQiskit").setLevel(logging.WARNING)
 
@@ -37,7 +37,7 @@ def split_string(s):
     return s.split(' ', 1)[0]
 
 
-class CustomQiskitNoisyBackend(Library):
+class CustomQiskitNoisyBackend(LibraryGenerative):
     """
     This module maps a library-agnostic gate sequence to a qiskit circuit and creates an artificial noise model.
     """
