+++ src/modules/applications/qml/generative_modeling/transformations/MinMax.py
@@ -16,8 +16,8 @@ from typing import Union
 import numpy as np
 
 from modules.applications.qml.generative_modeling.transformations.Transformation import Transformation
-from modules.circuits.CircuitStandard import CircuitStandard
-from modules.circuits.CircuitCardinality import CircuitCardinality
+from modules.applications.qml.generative_modeling.circuits.CircuitStandard import CircuitStandard
+from modules.applications.qml.generative_modeling.circuits.CircuitCardinality import CircuitCardinality
 
 
 class MinMax(Transformation):  # pylint: disable=R0902
