+++ src/modules/applications/qml/generative_modeling/transformations/PIT.py
@@ -16,7 +16,7 @@ import numpy as np
 import pandas as pd
 
 from modules.applications.qml.generative_modeling.transformations.Transformation import Transformation
-from modules.circuits.CircuitCopula import CircuitCopula
+from modules.applications.qml.generative_modeling.circuits.CircuitCopula import CircuitCopula
 
 
 class PIT(Transformation):  # pylint disable=R0902
